+++ .devel/sphinx/news.md
@@ -2,9 +2,13 @@
 
 ## 0.9.x (under development)
 
-*   The internal `Cdeadwood` function gained a new argument: `connected`.
-    It allows for choosing between chopping trees into chunks (default=false)
-    and branch pruning (true).
+*   [NEW FEATURE]  If the dataset is comprised of well-separated clusters
+    of heterogeneous densities, an attempt to split the dataset and refine
+    the outlierness markers is be made.
+
+*   [BACKWARD INCOMPATIBILITY] `kneedle_increasing` now returns
+    the index of the last element of a vector if a knee/elbow point could
+    not be identified.
 
 
 ## 0.9.0 (2026-02-13)